Clippy lints

+++ crates/ra_lsp_server/src/conv.rs
@@ -63,10 +63,7 @@ impl ConvWith for TextUnit {
 
     fn conv_with(self, line_index: &LineIndex) -> Position {
         let line_col = line_index.line_col(self);
-        Position::new(
-            u64::from(line_col.line),
-            u64::from(u32::from(line_col.col_utf16)),
-        )
+        Position::new(u64::from(line_col.line), u64::from(line_col.col_utf16))
     }
 }
 
@@ -204,10 +201,8 @@ impl TryConvWith for SourceChange {
                     .map(|it| it.edits.as_slice())
                     .unwrap_or(&[]);
                 let line_col = translate_offset_with_edit(&*line_index, pos.offset, edits);
-                let position = Position::new(
-                    u64::from(line_col.line),
-                    u64::from(u32::from(line_col.col_utf16)),
-                );
+                let position =
+                    Position::new(u64::from(line_col.line), u64::from(line_col.col_utf16));
                 Some(TextDocumentPositionParams {
                     text_document: TextDocumentIdentifier::new(pos.file_id.try_conv_with(world)?),
                     position,
